refactor: 공지사항 슬라이드 컴포넌트로 분리
